[Refactor] change directory
