+++ src/main/java/com/ada/earthvalley/yomojomo/activity/services/ActivityService.java
@@ -11,13 +11,13 @@ import com.ada.earthvalley.yomojomo.activity.dtos.ReadingListResponse;
 import com.ada.earthvalley.yomojomo.activity.dtos.ReadingListResponse.ReadingResponse;
 import com.ada.earthvalley.yomojomo.activity.entities.Activity;
 import com.ada.earthvalley.yomojomo.activity.entities.MainSentence;
-import com.ada.earthvalley.yomojomo.activity.repositories.ArticleRepository;
 import com.ada.earthvalley.yomojomo.article.entities.Article;
 import com.ada.earthvalley.yomojomo.article.entities.Topic;
 import com.ada.earthvalley.yomojomo.article.exceptions.ArticleError;
 import com.ada.earthvalley.yomojomo.article.exceptions.TopicError;
 import com.ada.earthvalley.yomojomo.article.exceptions.YomojomoArticleException;
 import com.ada.earthvalley.yomojomo.article.exceptions.YomojomoTopicException;
+import com.ada.earthvalley.yomojomo.article.repositories.ArticleRepository;
 import com.ada.earthvalley.yomojomo.article.repositories.TopicRepository;
 import com.ada.earthvalley.yomojomo.auth.SecurityUser;
 import com.ada.earthvalley.yomojomo.user.entities.User;
