+++ src/main/java/com/ada/earthvalley/yomojomo/activity/services/ReadingApiService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 
 import com.ada.earthvalley.yomojomo.activity.dtos.ReadingListResponse;
 import com.ada.earthvalley.yomojomo.activity.entities.Reading;
+import com.ada.earthvalley.yomojomo.article.services.ArticleDomainService;
 import com.ada.earthvalley.yomojomo.auth.SecurityUser;
 
 import lombok.RequiredArgsConstructor;
