+++ src/main/java/com/ada/earthvalley/yomojomo/article/services/ArticleDomainService.java
@@ -1,4 +1,4 @@
-package com.ada.earthvalley.yomojomo.activity.services;
+package com.ada.earthvalley.yomojomo.article.services;
 
 import static com.ada.earthvalley.yomojomo.article.exceptions.ArticleError.*;
 import static com.ada.earthvalley.yomojomo.article.exceptions.TopicError.*;
@@ -12,11 +12,11 @@ import org.springframework.transaction.annotation.Transactional;
 import com.ada.earthvalley.yomojomo.activity.dtos.ReadingListResponse;
 import com.ada.earthvalley.yomojomo.activity.dtos.ReadingListResponse.ReadingResponse;
 import com.ada.earthvalley.yomojomo.activity.entities.Reading;
-import com.ada.earthvalley.yomojomo.activity.repositories.ArticleRepository;
 import com.ada.earthvalley.yomojomo.article.entities.Article;
 import com.ada.earthvalley.yomojomo.article.entities.Topic;
 import com.ada.earthvalley.yomojomo.article.exceptions.YomojomoArticleException;
 import com.ada.earthvalley.yomojomo.article.exceptions.YomojomoTopicException;
+import com.ada.earthvalley.yomojomo.article.repositories.ArticleRepository;
 import com.ada.earthvalley.yomojomo.article.repositories.TopicRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -31,14 +31,13 @@ public class ArticleDomainService {
 	public List<ReadingResponse> articlesFromReadings(List<Reading> readings) {
 		return readings.stream().map(reading -> {
 
-				Article article = articleRepository.findById(reading.getArticleId())
-					.orElseThrow(() -> new YomojomoArticleException(ARTICLE_NOT_FOUND));
-				Topic topic = topicRepository.findById(article.getTopicId())
-					.orElseThrow(() -> new YomojomoTopicException(TOPIC_NOT_FOUND));
-				return new ReadingResponse(ReadingListResponse.TopicResponse.from(topic), article.getTitle(),
-					article.getId());
-			})
-			.collect(Collectors.toList());
+			Article article = articleRepository.findById(reading.getArticleId())
+				.orElseThrow(() -> new YomojomoArticleException(ARTICLE_NOT_FOUND));
+			Topic topic = topicRepository.findById(article.getTopicId())
+				.orElseThrow(() -> new YomojomoTopicException(TOPIC_NOT_FOUND));
+			return new ReadingResponse(ReadingListResponse.TopicResponse.from(topic), article.getTitle(),
+				article.getId());
+		}).collect(Collectors.toList());
 	}
 
 }
